Reorganized code structure | TODO integrate with scss and add show hide password button
